+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [0.2.3] - Unreleased
+**No Breaking Changes. Safe to Bump**
+### DirectorySummarizer Class
+- Updated to pass the max word length and max short summary length variables to the FileSummarizer.
+- Updated to only use the existing short summaries to generate the project contextual short summaries for each file.
+- Added functionality to fill in the blanks for files/directories still missing short summaries without including long summaries in the API call.
+- Added pagination for large directory structures using the new `DirectoryPaginator` class.
+- Skipped summarization for empty or near-empty files.
+
+### FileSummarizer Class
+- Modified to return both "summary" and "short_summary" in the same API call to reduce cost and response time.
+- Updated the prompt and API to return a JSON formatted response with the content summary applied to the key "summary" and the short summary to the key "short_summary".
+- Refactored code to validate this return value structure.
+- For larger content that is divided into chunks, the short summary is only generated on the final iteration where the summary is combined.
+
+### DirectoryPaginator Class
+- Created a new `DirectoryPaginator` class to handle pagination of large directory structures into smaller chunks.
+
+### Logger
+- Added more detailed `INFO` logs including directory size and files being summarized (optional argument).
+
+### Miscellaneous
+- Updated README with fixes.
+
 ## [0.2.2] - 2024-12-20
 **No Breaking Changes. Safe to Bump**
 ### DirectoryItem Class